+++ .github/workflows/mbt.yml
@@ -54,11 +54,17 @@ jobs:
         uses: actions-rust-lang/setup-rust-toolchain@v1
       - name: Install cargo-nextest
         uses: taiki-e/install-action@cargo-nextest
+      - name: Current time as random seed for Quint
+        run: echo "QUINT_SEED=$(date +%s)" >> $GITHUB_ENV
       - name: Build code
         working-directory: code/crates/test/mbt
         run: cargo nextest run -p informalsystems-malachitebft-test-mbt --all-features --no-run
-      - name: Current time as random seed for Quint
-        run: echo "QUINT_SEED=$(date +%s)" >> $GITHUB_ENV
       - name: Run tests
         working-directory: code/crates/test/mbt
         run: cargo nextest run -p informalsystems-malachitebft-test-mbt --all-features
+      - name: Build starknet code
+        working-directory: code/crates/starknet/test/mbt
+        run: cargo nextest run -p informalsystems-malachitebft-starknet-test-mbt --all-features --no-run
+      - name: Run starknet tests
+        working-directory: code/crates/starknet/test/mbt
+        run: cargo nextest run -p informalsystems-malachitebft-starknet-test-mbt --all-features
