+++ code/crates/starknet/test/mbt/specs
@@ -1,0 +1,1 @@
+../../../../../specs/starknet/
