+++ code/crates/starknet/test/mbt/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod deserializers;
+pub mod streaming;
+#[cfg(test)]
+pub mod tests;
