+++ code/crates/starknet/test/mbt/src/tests.rs
@@ -1,0 +1,1 @@
+pub mod streaming;
