+++ code/crates/starknet/test/mbt/src/tests/streaming.rs
@@ -1,8 +1,8 @@
 use glob::glob;
 use malachitebft_peer::PeerId;
+use malachitebft_test_mbt::utils::{generate_random_traces, generate_test_traces, quint_seed};
 
 use crate::streaming::State;
-use crate::utils::{generate_random_traces, generate_test_traces, quint_seed};
 
 pub mod runner;
 pub mod utils;
@@ -22,6 +22,7 @@ fn test_mbt_part_streaming_specified_traces() {
     let quint_seed = quint_seed();
 
     print!("{}\n", temp_path.to_string_lossy());
+    // WARNING: After moving starknet related specs to separate repo this probably won't  work
     generate_test_traces(
         "starknet/block-streaming/part_stream.qnt",
         &temp_path.to_string_lossy(),
@@ -61,6 +62,7 @@ fn test_mbt_part_streaming_random_traces() {
     let quint_seed = quint_seed();
 
     print!("{}\n", temp_path.to_string_lossy());
+    // WARNING: After moving starknet related specs to separate repo this probably won't  work
     generate_random_traces(
         "starknet/block-streaming/part_stream.qnt",
         &temp_path.to_string_lossy(),
