+++ code/crates/test/mbt/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod consensus;
 pub mod deserializers;
-pub mod streaming;
 pub mod types;
 pub mod utils;
 pub mod votekeeper;
