+++ code/crates/test/mbt/src/tests.rs
@@ -1,3 +1,2 @@
 pub mod consensus;
-pub mod streaming;
 pub mod votekeeper;
